+++ pkg/log/log.go
@@ -47,23 +47,3 @@ func (z *logger) Errorw(msg string, keysAndValues ...interface{}) {
 func (z *logger) Fatalw(msg string, keysAndValues ...interface{}) {
 	syslog.Fatal(msg, keysAndValues)
 }
-
-func Debugw(msg string, keysAndValues ...interface{}) {
-	Default.Debugw(msg, keysAndValues...)
-}
-
-func Infow(msg string, keysAndValues ...interface{}) {
-	Default.Infow(msg, keysAndValues...)
-}
-
-func Warnw(msg string, keysAndValues ...interface{}) {
-	Default.Warnw(msg, keysAndValues...)
-}
-
-func Errorw(msg string, keysAndValues ...interface{}) {
-	Default.Errorw(msg, keysAndValues...)
-}
-
-func Fatalw(msg string, keysAndValues ...interface{}) {
-	Default.Fatalw(msg, keysAndValues...)
-}
