+++ templates/clients.html
@@ -28,6 +28,7 @@
           </tr>
           {{ end }}
       </table>
+
 <h4>Creating new client:</h4>
 <div class="row">
 <div class="col-xs-10 col-md-6">
